remove lblockindex and conf.cfg.DataDir
