+++ logic/lreindex/lreindex_test.go
@@ -44,11 +44,9 @@ func initTestEnv(t *testing.T) (dirpath string, err error) {
 		model.SetRegTestParams()
 	}
 
-	persist.InitPersistGlobal()
-
 	// Init UTXO DB
 	utxoDbCfg := &db.DBOption{
-		FilePath:  conf.Cfg.DataDir + "/chainstate",
+		FilePath:  conf.DataDir + "/chainstate",
 		CacheSize: (1 << 20) * 8,
 		Wipe:      conf.Cfg.Reindex,
 	}
@@ -57,17 +55,16 @@ func initTestEnv(t *testing.T) (dirpath string, err error) {
 
 	// Init blocktree DB
 	blkDbCfg := &db.DBOption{
-		FilePath:  conf.Cfg.DataDir + "/blocks/index",
+		FilePath:  conf.DataDir + "/blocks/index",
 		CacheSize: (1 << 20) * 8,
 		Wipe:      conf.Cfg.Reindex,
 	}
 	blkdbCfg := blkdb.BlockTreeDBConfig{Do: blkDbCfg}
 	blkdb.InitBlockTreeDB(&blkdbCfg)
 
-	chain.InitGlobalChain()
+	chain.InitGlobalChain(blkdb.GetInstance())
 
-	// Load blockindex DB
-	lblockindex.LoadBlockIndexDB()
+	persist.InitPersistGlobal(blkdb.GetInstance())
 
 	mempool.InitMempool()
