+++ model/utxo/coindb_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/copernet/copernicus/model/outpoint"
 	"github.com/copernet/copernicus/model/script"
 	"github.com/copernet/copernicus/model/txout"
+	"github.com/copernet/copernicus/persist/blkdb"
 	"github.com/copernet/copernicus/persist/db"
 	"github.com/copernet/copernicus/util"
 	"github.com/copernet/copernicus/util/amount"
@@ -40,7 +41,7 @@ func TestCoinsDB(t *testing.T) {
 		t.Errorf("dbw get by GetDBW is not equal to obj.dbw")
 	}
 
-	chain.InitGlobalChain()
+	chain.InitGlobalChain(blkdb.GetInstance())
 
 	if err != nil {
 		t.Fatal(err)
