+++ service/blockservice_test.go
@@ -7,7 +7,6 @@ import (
 	"github.com/copernet/copernicus/conf"
 	"github.com/copernet/copernicus/crypto"
 	"github.com/copernet/copernicus/log"
-	"github.com/copernet/copernicus/logic/lblockindex"
 	"github.com/copernet/copernicus/logic/lchain"
 	"github.com/copernet/copernicus/logic/ltx"
 	"github.com/copernet/copernicus/model"
@@ -65,11 +64,9 @@ func initTestEnv(t *testing.T, args []string, initScriptVerify bool) (dirpath st
 	}
 	log.Init(string(configuration))
 
-	persist.InitPersistGlobal()
-
 	// Init UTXO DB
 	utxoDbCfg := &db.DBOption{
-		FilePath:  conf.Cfg.DataDir + "/chainstate",
+		FilePath:  conf.DataDir + "/chainstate",
 		CacheSize: (1 << 20) * 8,
 		Wipe:      conf.Cfg.Reindex,
 	}
@@ -78,17 +75,15 @@ func initTestEnv(t *testing.T, args []string, initScriptVerify bool) (dirpath st
 
 	// Init blocktree DB
 	blkDbCfg := &db.DBOption{
-		FilePath:  conf.Cfg.DataDir + "/blocks/index",
+		FilePath:  conf.DataDir + "/blocks/index",
 		CacheSize: (1 << 20) * 8,
 		Wipe:      conf.Cfg.Reindex,
 	}
 	blkdbCfg := blkdb.BlockTreeDBConfig{Do: blkDbCfg}
 	blkdb.InitBlockTreeDB(&blkdbCfg)
 
-	chain.InitGlobalChain()
-
-	// Load blockindex DB
-	lblockindex.LoadBlockIndexDB()
+	chain.InitGlobalChain(blkdb.GetInstance())
+	persist.InitPersistGlobal(blkdb.GetInstance())
 
 	err = lchain.InitGenesisChain()
 	assert.Nil(t, err)
